Remove printlns
These prints seem to just be for debugging and print values returned by
the functions. Removing them would be better for users of the crate
cleaning up their programs output as well as providing a minor
performance improvement (this provided a 1ms improvement per iteration
for me ~28% improvement on average.